+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed incorrect color conversions in the color picker between RGB/HSV/Hex
 - Fixed autoimporting modules not respecting multiline requires
 - Fixed documentation for `debug` library
+- Fixed synthetic `typeof()` showing up in signature help for builtin tables (e.g. `function typeof(string).byte(...)`)
 
 ## [1.21.0] - 2023-06-14